% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFunnel.R
\name{saveFunnel}
\alias{saveFunnel}
\title{Save a funnel plot to an HTML file}
\usage{
saveFunnel(plot, file, selfcontained = TRUE)
}
\arguments{
\item{plot}{Plot to save}

\item{file}{File to save HTML into}

\item{selfcontained}{Whether to save the HTML as a single self-contained file
(with external resources base64 encoded) or a file with external resources
placed in an adjacent directory.}
}
\description{
Save a funnel plot to an HTML file
}

