% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeyBarChart.R
\name{sankeyBarChart}
\alias{sankeyBarChart}
\title{Create a D3 JavaScript Sankey Barchart}
\usage{
sankeyBarChart(data, height = NULL, width = NULL)
}
\arguments{
\item{data}{A data frame with two column: label - contains the name of steps,
value - contain the value of the steps}

\item{height}{height for the plot (if \code{NULL} then width is automatically
determined based on context.)}

\item{width}{width for the plot (If \code{NULL} then height is automatically
determined based on context, allowing reactivity)}
}
\description{
To create a google-analytics-like sankey barchart.
}
\examples{
\dontrun{
   data <- data.frame(label=c("step 1","step 2","step 3","step 4"),
                      value=c(1000,600,400,100))
   sankeyBarChart(data)
}
}

